Categorias e Carteiras

+++ nanas-backend/src/main/java/com/nanas/nanas/model/Usuario.java
@@ -1,6 +1,6 @@
 package com.nanas.nanas.model;
 
-import jakarta.persistence.*;
+import javax.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
@@ -13,7 +13,7 @@ import lombok.AllArgsConstructor;
 public class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String email;
     private String nome;
     private String firebaseUid;
