+++ nanas-backend/src/main/java/com/nanas/nanas/repository/TransacaoRepository.java
@@ -5,7 +5,7 @@ import com.nanas.nanas.model.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface TransacaoRepository extends JpaRepository<Transacao, Integer> {
+public interface TransacaoRepository extends JpaRepository<Transacao, Long> {
     List<Transacao> findByUsuarioOrderByDataDesc(Usuario usuario);
 }
